cdks/requirements.txt
psycopg2==2.9.10
aws-cdk-lib==2.196.0
constructs==10.4.2
